Download KNMI hourly data to file

Meer info, zie https://www.knmi.nl/kennis-en-datacentrum/achtergrond/data-ophalen-vanuit-een-script

download_knmi_data_hour(stations, variables, start_date, end_date,
  output_file = "knmi_download.txt")

Arguments

stations
list of integers, for an overview, see http://projects.knmi.nl/klimatologie/metadata/stationslijst.html

variables
list of variables, options are:

"WIND = DD:FH:FF:FX"Wind

"TEMP = T:T10N:TD"Temperatuur

"SUNR = SQ:Q"Zonneschijnduur en globale straling

"PRCP = DR:RH"Neerslag en potentiële verdamping

"VICL = VV:N:U"Zicht, bewolking en relatieve vochtigheid

"WEER = M:R:S:O:Y:WW"Weerverschijnselen, weertypen

"ALL"alle variabelen

start_date
date interpretable string to define start of the period

end_date
date interpretable string to define end of the period

output_file
path/filename of the output_file_name

Value

response containing, status_code, content,...

Examples

# NOT RUN {
download_knmi_data_uur(c(310, 319), "PRCP", "2011-01-01", "2012-02-17",
output_file = "knmi_output_download.txt")
# }