Take a dataset and plot in different projections on a leaflet map to check the expected projection

When retrieving a number of data-points, but the kind of projection is not provided/known (for any reason), this utility plots the data for different projections on a map to make comparison possible. The main function is guess_projection which creates the map with different options. Custom projections can be provided by the user as well.

guess_projection(
  df,
  col_long,
  col_lat,
  belgium = TRUE,
  projections = c("epsg:4326", "epsg:31370", "epsg:28992", "epsg:32631", "epsg:3812",
    "epsg:3035")
)

Arguments

df

data.frame with a x and y coordinate column

col_long

(char) name of the x (longitude) column

col_lat

(char) name of the y (latitude) column

belgium

(bool) If TRUE, coordinates are expected to be in Belgium

projections

(list) epsg codes of the different projections to evaluate By default, the following six projections are added to the map:

"epsg:4326"WGS 84, http://spatialreference.org/ref/epsg/wgs-84/

"epsg:31370"Belge 1972/Belgian Lambert 72, http://spatialreference.org/ref/epsg/31370/

"epsg:28992"Amersfoort/Rijksdriehoek nieuw, http://spatialreference.org/ref/epsg/28992/

"epsg:32631"WGS 84 / UTM zone 31N, http://spatialreference.org/ref/epsg/32631/

"epsg:3812"ETRS89/Belgian Lambert 2008, http://spatialreference.org/ref/epsg/3812/

"epsg:3035"ETRS89 / ETRS-LAEA, http://spatialreference.org/ref/epsg/3035/

Details

For each of the given projections, the coordinates columns are given the the specified projection. In a next step, these projections are converted to wgs84 and plotted on a openstreetmap background with leaflet. The interactive map provides the possibility to select/unselect specific layers.

See also

Examples

if (FALSE) {
guess_projection(data, "x", "y")
}