Reproject x-y coordinates from data.frame columns.

This function extracts x-y coordinates from a data.frame by means of the given coordinate reference system (CRS), converts them to the new CRS and assign them back to the data.frame columns.

reproject_coordinates(df, col_long, col_lat, crs_input, crs_output)

Arguments

df
A data.frame with a x and y coordinate column.

col_long, col_lat
Column names or positions of the x (longitude) and y (latitude) column. They are passed to vars_pull. These arguments are passed by expression and support quasiquotation (you can unquote column names or column positions).

crs_input
Projection string of class CRS-class (sp compatible) or crs-class (sf compatible) defining the current projection.

crs_output
Projection string of class CRS-class (sp compatible) or crs-class (sf compatible) defining the projection to convert to.

Value

A data.frame with the same columns, but adapted coordinates for the x and y column values.

See also

Examples

library(sp)
library(sf)

#> Linking to GEOS 3.8.1, GDAL 3.1.4, PROJ 6.3.1

data_pts <- data.frame(
  id = c(1, 2),
  lat = c(51.23031, 50.76931),
  lon = c(5.083980, 3.829593),
  stringsAsFactors = FALSE
)

# CRS-class (use sp package)
sp_crs1 <- CRS("+init=epsg:4269")
sp_crs2 <- CRS("+init=epsg:3857")

#> Warning: Discarded ellps WGS 84 in Proj4 definition: +proj=merc +a=6378137 +b=6378137 +lat_ts=0 +lon_0=0 +x_0=0 +y_0=0 +k=1 +units=m +nadgrids=@null +wktext +no_defs

#> Warning: Discarded datum WGS_1984 in Proj4 definition

reproject_coordinates(data_pts, col_long = "lon", col_lat = "lat",
  crs_input = sp_crs1, crs_output = sp_crs2)

#>   id     lat      lon
#> 1  1 6662134 565946.1
#> 2  2 6580588 426308.3

# crs-class (use sf package)
sf_crs1 <- st_crs(4269)
sf_crs2 <- st_crs(3857)
reproject_coordinates(data_pts, col_long = "lon", col_lat = "lat",
  crs_input = sf_crs1, crs_output = sf_crs2)

#>   id     lat      lon
#> 1  1 6662134 565946.1
#> 2  2 6580588 426308.3

# input projection is CRS-class (sp) and output projection crs-class (sf)
reproject_coordinates(data_pts, col_long = "lon", col_lat = "lat",
  crs_input = sp_crs1, crs_output = sf_crs2)

#>   id     lat      lon
#> 1  1 6662134 565946.1
#> 2  2 6580588 426308.3

# input projection is crs-class (spf and output projection CRS-class (sp)
reproject_coordinates(data_pts, col_long = "lon", col_lat = "lat",
  crs_input = sf_crs1, crs_output = sp_crs2)

#>   id     lat      lon
#> 1  1 6662134 565946.1
#> 2  2 6580588 426308.3

# use names (character) of x-y columns
reproject_coordinates(data_pts, col_long = "lon", col_lat = "lat",
  crs_input = sf_crs1, crs_output = sf_crs2)

#>   id     lat      lon
#> 1  1 6662134 565946.1
#> 2  2 6580588 426308.3

# use NSE of x-y columns
reproject_coordinates(data_pts, col_long = lon, col_lat = lat,
  crs_input = sf_crs1, crs_output = sf_crs2)

#>   id     lat      lon
#> 1  1 6662134 565946.1
#> 2  2 6580588 426308.3

# use position of x-y columns
reproject_coordinates(data_pts, col_long = 3, col_lat = 2,
  crs_input = sf_crs1, crs_output = sf_crs2)

#>   id     lat      lon
#> 1  1 6662134 565946.1
#> 2  2 6580588 426308.3