Variance inflation factor

Variance inflation factor

Variance inflation factor (lm model)

Variance inflation factor (data.frame)

Variance inflation factor (default routine)

vif(mod, ...)

# S3 method for lm
vif(mod, ...)

# S3 method for data.frame
vif(mod, ...)

# S3 method for default
vif(mod, ...)

Arguments

mod

A model object (or data.frame). For the moment only lm and a plain dataset is implemented

...

not used

Value

a matrix with for each variable de variance inflation factor, the degrees of freedom en the rescaled variance inflation factor based on the degrees of freedom.

Examples

{
vif(airquality)
}
#> Warning: essentially perfect fit: summary may be unreliable
#>             GVIF Df GVIF^(1/(2*Df))
#> Ozone   2.666246  1        1.632864
#> Solar.R 1.202654  1        1.096656
#> Wind    1.664939  1        1.290325
#> Temp    2.508438  1        1.583805
#> Month   1.305582  1        1.142621
#> Day     1.024798  1        1.012323