Add species information provided by the GBIF taxonomic backbone API

This functions extends an existing dataframe with additional columns provided by the GBIF taxonomic backbone and matched on the species scientific name, which need to be an available column in the dataframe.

gbif_species_name_match(df, name_col, gbif_terms = c("usageKey",
  "scientificName", "rank", "order", "matchType", "phylum", "kingdom", "genus",
  "class", "confidence", "synonym", "status", "family"))

Arguments

df
data.frame with species information

name_col
char column name of the column containing the scientific names used for the name matching with the GBIF taxonomic backbone.

gbif_terms
list of valid GBIF terms to add as additional columns to the data.frame

Value

df with GBIF information as additional columns

Details

This function is essentially a wrapper around the existing rgbif functionality `name_lookup()` and extends the application to a data.frame. For more information on the name matching API of GBIF on which rgbif relies, see https://www.gbif.org/developer/species