inborutils: Collection of Useful R Utilities

While working on research projects, typical small functionalities are useful across these projects. Instead of copy-pasting these functions in all individual project repositories/folders, this package collects these functions for reuse by ourself and - if useful - others as well.

See also

Author

Maintainer: Hans Van Calster hans.vancalster@inbo.be (ORCID) (Research Institute for Nature and Forest (INBO))

Authors:

Damiano Oldoni damiano.oldoni@inbo.be (ORCID) (Research Institute for Nature and Forest (INBO))

Stijn Van Hoey stijnvanhoey@gmail.com (ORCID) (Research Institute for Nature and Forest (INBO))

Other contributors:

Dimitri Brosens dimitri.brosens@inbo.be (ORCID) (Research Institute for Nature and Forest (INBO)) [contributor]

Peter Desmet peter.desmet@inbo.be (ORCID) (Research Institute for Nature and Forest (INBO)) [contributor]

Frank Huysentruyt frank.huysentruyt@inbo.be (ORCID) (Research Institute for Nature and Forest (INBO)) [contributor]

Thierry Onkelinx thierry.onkelinx@inbo.be (ORCID) (Research Institute for Nature and Forest (INBO)) [contributor]

Floris Vanderhaeghe floris.vanderhaeghe@inbo.be (ORCID) (Research Institute for Nature and Forest (INBO)) [contributor]

Pieter Verschelde pieter.verschelde@inbo.be (ORCID) (Research Institute for Nature and Forest (INBO)) [contributor]

Els De Bie els.debie@inbo.be (ORCID) (Research Institute for Nature and Forest (INBO)) [contributor]

Nicolas Noé nicolas.noe@inbo.be (ORCID) (Research Institute for Nature and Forest (INBO)) [contributor]

Els Lommelen els.lommelen@inbo.be (ORCID) (Research Institute for Nature and Forest (INBO)) [contributor]

Research Institute for Nature and Forest (INBO) info@inbo.be [copyright holder, funder]